fix_naive.cpp

+++ data_structure/static_range_sum_with_upper_bound/sol/naive.cpp
@@ -6,8 +6,8 @@ int main(){
     int n,q;
     scanf("%d %d",&n,&q);
 
-    vector<int> a(n);
-    for(int i = 0;i < n;i++) scanf("%d",&a[i]);
+    vector<long long> a(n);
+    for(int i = 0;i < n;i++) scanf("%lld",&a[i]);
 
     for(int query = 0;query < q;query++){
         int t,l,r,x;
